+++ src/Models/Cart.php
@@ -18,6 +18,8 @@ use Konekt\Enum\Eloquent\CastsEnums;
 use Vanilo\Cart\Exceptions\InvalidCartConfigurationException;
 use Vanilo\Contracts\Buyable;
 use Vanilo\Cart\Contracts\Cart as CartContract;
+use Vanilo\Contracts\Decimal as DecimalContract;
+use Vanilo\Support\Decimal;
 
 class Cart extends Model implements CartContract
 {
@@ -115,9 +117,9 @@ class Cart extends Model implements CartContract
     /**
      * @inheritDoc
      */
-    public function total()
+    public function total(): DecimalContract
     {
-        return $this->items->sum('total');
+        return new Decimal($this->items->sum('total'));
     }
 
     /**