+++ tests/Dummies/SizedProduct.php
@@ -21,4 +21,9 @@ class SizedProduct extends Model implements Buyable
     use BuyableModel, BuyableNoImage;
 
     protected $guarded = ['id'];
+
+    public function getCurrency(): string
+    {
+        return 'USD';
+    }
 }